Test dual-GPU path for high-performance WebGL contexts on Mac/Metal.
The flag EGLDualGPURendering enables the dual-GPU path to be used.
This flag will be removed once this is well tested.
The flag ForceHighPerformanceGPUForWebGL forces all WebGL contexts to
use the high performance GPU, even if the rest of Chrome's rendering
is on the low power GPU. This is different from using
--force_high_performace_gpu which forces all of Chrome's rendering to
the high performance GPU.

Bug: chromium:1251724
Change-Id: Ifbc4c47c60e567388ad073b580aba05d4f28575a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3978571
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1106227}
NOKEYCHECK=True
GitOrigin-RevId: 16f60080d5f2e836a73f3dcf36a753749c00fb80