Rename "libfuzzer_main" to "fuzzer_engine_main".
This gn target was misnamed, because it already applied to multiple
fuzzing engines - AFL as well as libfuzzer. This CL addresses a
longstanding TODO to rename the target name to be more accurate.

In the future this will become more important as this target is used not
just for libfuzzer and AFL but also for centipede.

It also addresses some iOS filenames which are unnecessarily
libfuzzer-specific.

This CL should have no functional changes.

Bug: 1369919
Change-Id: I02969b4e32700d717b679ca1ad426cfb5cdfd589
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4286263
Reviewed-by: Chris Bookholt <bookholt@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1109654}
NOKEYCHECK=True
GitOrigin-RevId: f259aa82264b21bcceaf73acfc9652c7cf57e34c