Switch to simpler base::WriteFile() variants (132/N)
This CL is part of a batch of CLs to replace the use of the less
friendly base::WriteFile call with simpler variants, which are easier to
read, and less prone to mistakes.

Changes in this particular CL address files under the path
/testing/data_driven_testing.

This CL was uploaded by git cl split.

R=fleimgruber@google.com

AX-Relnotes: n/a.
Bug: 418837
Change-Id: Ib1bcca01ad3a2df86b0e97720446c032a52f9c8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4292115
Auto-Submit: Claudio DeSouza <cdesouza@igalia.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1110222}
NOKEYCHECK=True
GitOrigin-RevId: 94ad878fe8a5e79075e3f02b685eddface3928e9