Feature to reorder touch and motion handling.
The speed at which touch events reach the renderer is latency
sensitive. We should avoid doing any significant work. We believe
that the observers of onMotion are currently taking a measurable
amount of time in the suppression experiment.

To mitigate this problem, this CL adds a feature that reorders these
two operations. First the touch event is sent to children via dispatch
and it eventually reaches native/renderer. And then second the
inMotion supplier is updated.

Bug: 1419567
Change-Id: I57d2ae9f66097b714f74b9851e31582c3e0bd258
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4292042
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Sky Malice <skym@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1110594}
NOKEYCHECK=True
GitOrigin-RevId: d4ef5106d9bfe09fa9e861d11da6473a629ad67f