Fix UPF stale comment removal
Fixes a bug in the unexpected pass finder that caused it to mishandle
stale comment removal calculations due to improperly keeping track of
which lines were removed. In the unittests, this manifested as an index
out of bounds exception, but it is believed that this is also the root
cause for a weird comment removal we saw in the wild.

Also re-adds the part of the comment that was erroneously removed in
the CL that caused us to notice this issue.

Bug: 1395078
Change-Id: I42ce0b80e7f1136f4fcc7456defbaf9564d455dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4363360
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1121201}
NOKEYCHECK=True
GitOrigin-RevId: c42fd43d49594c08cb1ce10031b0aae0f5005122