Remove base::Feature MojoAvoidRandomPipeId
Set the behavior of base::Feature MojoAvoidRandomPipeId as default and
remove the feature. The behavior was found to be slightly useful. From
an internal document:

  Sometimes better battery usage (0.6%). Observing some extra janky
  intervals (0.1%), possibly from the background.

MojoAvoidRandomPipeId was folded into the field trial
AndroidScrollOptimizations to evaluate it with other optimizations as a
collection. Due to problems rolling out that collection, rolling this
out as a default instead.

Bug: 1351813
Change-Id: Ia3a8f5606f1c681116ab615cda53a6efa507575a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4373638
Commit-Queue: Egor Pasko <pasko@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/main@{#1122471}
NOKEYCHECK=True
GitOrigin-RevId: e3c2374e2bacbb9b385259df82323177b7f16073