[Sheriff] Fix test in TabbedAppMenuTest and renable in Android 12
Fixed testQuickDeleteMenu_NotShownInIncognito in the test suite which
was doing a menu assertion on regular menu in some case, which gets
opened up default in the setup method.

The fix is to hide the menu in the test first and then continue the
execution.

Bug: 1429722
Change-Id: I92ccc35fb02c0b0356ccda68e8d1736eb30dcc80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4402259
Reviewed-by: Ehimare Okoyomon <eokoyomon@chromium.org>
Owners-Override: Rohit Agarwal <roagarwal@chromium.org>
Commit-Queue: Rohit Agarwal <roagarwal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1126530}
NOKEYCHECK=True
GitOrigin-RevId: bf221b2195cef2cce9c968ee80bbe8b79c8c9bcc