Add field trial for WebApkInstallFailureAndRetry
AppBannerManagerTest.java:
These two tests were meant for testing Installs (WebAPK) but checking
if the splash screen icon was downloaded, that's for legacy A2HS.

AddToHomescreenTest.java:
The test was for legacy A2HS, use a manifest with no name so that
it triggers the A2HS flow instead of the install one.
It (and most tests in this file) will need to be updated once universal
install is implemented.

Bug: 1409642
Change-Id: I9ffdd49246871f45083bb23fa5d8904ae7c551a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4391931
Reviewed-by: Glenn Hartmann <hartmanng@chromium.org>
Commit-Queue: Ella Ge <eirage@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1126753}
NOKEYCHECK=True
GitOrigin-RevId: d7b5a79a22427bb2b435b051a5debb42cf92218a