Modify generate_buildbot_json.py to use absolute paths.
generate_buildbot_json passes around many filenames which may be
absolute paths if the pyl files dir is overridden by a flag or a simple
file name. This requires all of the code working with the files to
account for this and results in repeated calculation of the file paths
and makes changing the location of gn_isolate_map.pyl difficult.

This changes it so that absolute file paths are passed around instead,
which consolidates the absolute path calculation into only a couple
spots. As part of this, an --output-dir flag has been added so that an
absolute output directory location is available to the code.

Bug: 1426635
Change-Id: Icb51febf2eb412261664d325d7090470827e9d61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4409891
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1128313}
NOKEYCHECK=True
GitOrigin-RevId: 79339e1810f9ca6c6f295265d95ae9fb7415e9d8