[RenderText] Add Crash reporting to debug fallback fonts.
We need to determine why we are still getting cases that fallback
to GetFallbackFonts code. We gate DumpWithoutCrashing with an experiment
that will be turned on only for Canary and Dev on Windows. We will not
be getting crash information on Beta or Stable.

Set the following in crash keys:
- Primary font name
- Primary font script

Save the following on the stack to be analyzed through minidump
- Text passed into ShapeRuns
- Runs that we shaped with a font from GetFallbackFonts
- font name used to shape runs in GetFallbackFonts

Bug: 995789
Change-Id: I3989de3964492b9edd7c0154129a10bc40f88c93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4310524
Commit-Queue: Elaine Chien <elainechien@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128389}
NOKEYCHECK=True
GitOrigin-RevId: f983756ec1cf22ed3c87e929df38a7d22a98db92