Fix generate_buildbot_json_unittest.py to pass on Windows.
https://crrev.com/c/4409891 modified generate_buildbot_json.py to work
with absolute paths. As part of that change, the test uses these
absolute paths, including in regular expressions on exception messages.
This breaks on Windows because the paths contain backslash characters
that the regex engine interprets as regular expression escape sequences.
This changes it so that these paths get escaped before being included in
a regular expression.

Bug: 1432281
Change-Id: I6d6b2cba86a4ad5e64d1f984b7b511b6f3929e2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4416380
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1129028}
NOKEYCHECK=True
GitOrigin-RevId: 807011ab2fe15e07528cd84144702926c80bf0c5