Fix context menu browser test on Lacros.
The browser_test ContextMenuBrowserTest.OpenLinkInProfile depends on
creating a bunch of browser windows with different profiles for them to
be counted as "active." Lacros seems to make this more asynchronous
which makes this fail more reliably, but after debugging I believe this
call is technically required on all platforms for the test to be
reliable.

This waits for the browser windows to open for the OpenLinkInProfile
test which fixes it on Lacros.

I believe this test can be enabled on all platforms now (it's still
disabled in some configurations) but I'll enable those in a followup.

Bug: 1150200
Change-Id: I54f8b03ef7d24b1106605cd801fdc1f92f9c691b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4400979
Commit-Queue: Brett Wilson <brettw@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1129581}
NOKEYCHECK=True
GitOrigin-RevId: dd01bd7df9057a4928b84442d059e0b704d94824