chromeOS: Split existing memory pressure experiment in 2.
This CL is not intended to have any immediate effect as it splits a
disabled-by-default experiment into 2 disabled-by-default experiments.

We want to change the behavior of the existing
ChromeOSMemoryPressureSignal experiment to have different behavior for
ARC-enabled and ARC-disabled devices. This CL splits the experiment in 2
and updates the fieldtrial_testing_config.

This CL uses arc::IsArcAvailable as a proxy for whether arc is enabled.
This is safe to rely on in early lifecycle since the function only
relies on some command line flag checks. This means that some users who
are not actually using ARC will end up in the arc-enabled experiment.
This is expected and appropriate since the arc-enabled experiment is
safer and less aggressive than the arc-disabled experiment.

Change-Id: I7ce06e3b66b332a7aebf1a150b51704715b651d1
Bug: b/226425011
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4455836
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1134907}
NOKEYCHECK=True
GitOrigin-RevId: bbc884ec2519cdedf6ff7036b28c741f4ff2e5e5