Fix omnibox test for Lacros.
OmniboxPopupViewViewsTest.ClickOmnibox failed on Lacros. It works by
setting up an omnibox selection and sending a click, expecting that a
handled click will clear the selection.

The problem is that if the omnibox doesn't have focus, the click will
automatically select all text which will confuse the test. To avoid the
automatic select all and remove the dependency on platform docus
handling, this instead sends a drag event to the omnibox.

I believe this is likely to fix Mac where it is disabled, but will
check in a followup.

Change-Id: If8600a54c1fbc02bb47baffd7f8f1a2a4cf86d8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4476664
Reviewed-by: Khalid Peer <khalidpeer@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1135608}
NOKEYCHECK=True
GitOrigin-RevId: 47c617325c3d81e225db9da9143ff146a4e31c38