Fix VariationsSafeModeEndToEndBrowserTest on official builds.
Certain platform differences make using either an abort() or
a LOG(FATAL) problematic:

On Windows, in release/official, an abort generates a fastfail and
in certain configurations, Windows Error Reporting can capture
these fastfails and attempting to create a dump of browser_tests.exe,
which can cause the test to fail or hang.

On iOS, a LOG(FATAL) does not always abort, so an abort() was needed.

In order to avoid these platform specific complexities, the test
is changed to instead terminate the process with a custom
exit code, which is checked in the test.

This makes the test reliable on all platform configurations.

BUG=1400987,1344852,1297123,1326982
Include-Ci-Only-Tests: true
Disable-Rts: true

Change-Id: Ibb68f0b0dcd1fd9e1b2a707a8ce57b9385b7a5d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4486398
Reviewed-by: Ginny Huang <ginnyhuang@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Roger McFarlane <rogerm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1139035}
NOKEYCHECK=True
GitOrigin-RevId: d99729116120c19afa30c28e3c02eb082c4f2f8a