Enable more bluetooth extension APIs for Lacros.
Enables the bluetoothPrivate and bluetoothLowEnergy extension APIs for
Lacros. Previously these were only enabled for the Ash browser
(extensions platform = "chromeos").

Enables the BluetoothLowEnergyApiTest.* and BluetoothPrivateApiTest.*
tests which now pass.

Also enables the feedbackPrivate.readLogSource extension API on Lacros. This seems to be not covered by a test but I believe should not differ between the Ash Browser and Lacros. This "API is restricted to allowlisted extensions and WebUI." Currently WebUI on ChromeOS is still run in the Ash browser which is probably why we have not noticed, but this is likely to change when these use-cases are converted to Lacros.

Bug: 1111979
Include-Ci-Only-Tests: true
Change-Id: I3bd3e4f9e5aa72fff411fb820120008f61d92f22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4500344
Commit-Queue: Brett Wilson <brettw@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141050}
NOKEYCHECK=True
GitOrigin-RevId: 6c6b786107d38d259fbeea5159fecda2cfec8d0d