Enable EsbIphBubbleAndCollapseSettings and parameters by default
The launch for this feature has been approved so we are enabling it at
the tip-of-tree. After this changelist, when users click on the enhanced
protection promotion on a security interstitial, they will see an
in-product-help bubble and the enhanced protection radio button will be
collapsed.

This change also removes the associated entry from
fieldtrial_testing_config.json

Bug: 1443626
Change-Id: Ic4a4f2b6120d618f187e15c3bc73cc60cc3fd54e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4515457
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Javier Castro <jacastro@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141603}
NOKEYCHECK=True
GitOrigin-RevId: 32c177a844bfe9eda8d8fac5bdc4fdc37b5f2c17