Reland "[dPWA] Add OsIntegrationSubManagers to field trial testing"
This is a reland of commit a167246101099d3fe7fcdc84af27cdde2ffa00d6

This CL fixes the Mac failures on the revert build by ensuring that
AppShimRegistry information is created for an app and profile as
part of the ShortcutSubManager::Configure() flow. This is necessary
for the FileHandlingSubManager calls to the AppShimRegistry inside
its Configure() flow to execute, otherwise the correct data is not
written to the AppShimRegistry. Previously, the ShortcutSubManager
would only create the AppShimRegistry on new installs on Execute(),
and the FileHandlingSubManager would not correctly store that
information.

Protocol Handling follows a similar path but it is not an issue
because AppShimRegistry is touched only as part of the Execute flow,
and it is always called AFTER ShortcutSubManager::Execute, so the
AppShimRegistry is always created per app_id and profile.

Duplicate calls to AppShimRegistry::OnAppInstalledForProfile() is
fine, if the registry is already set up fr that profile and app_id,
the function is a no-op and exits early.

Tested by running all the unit-tests for these 3 sub managers, on
all Mac trybots from Mac 11-13 and the browsertests that failed on
https://ci.chromium.org/p/chromium/builders/ci/Mac12%20Tests/9259.

Original change's description:
> [dPWA] Add OsIntegrationSubManagers to field trial testing
>
> Bug: b/277775035
> Change-Id: I9d8c3d7b2f43481825bc951579ba3bf7a07ff4c8
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4447618
> Reviewed-by: Daniel Murphy <dmurph@chromium.org>
> Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1141697}

Bug: b/277775035
Change-Id: I79254fbe847aec9bad67cde851cda0a6729022a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4517975
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1142926}
NOKEYCHECK=True
GitOrigin-RevId: 195df62b68fa1798379a91f09ac0c695388c8e3a