De-flake PlatformAppHasDifferentAudioFocus
The std::list containing the app windows created in this test are in the
"opposite" order about 15% of the time on my system. When this happens,
it makes the assumptions the test makes invalid.

But the test doesn't care about the order of the windows, it's just
checking that the two created windows have different audio contexts.
Updates the test to allow the windows to be stored in either order.

Bug: 1352629, 1352523, b/262651068
Change-Id: Ib3e17db29d26b8bc05aab8557a48dea8ec7d07c3
Include-Ci-Only-Tests: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4553682
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1147563}
NOKEYCHECK=True
GitOrigin-RevId: 385f32dd8a759732b0edf475a501838afdcc482e