[rust] Accept static initializer for std::os::args
Rust includes a static initializer in its standard library, which is
used to populate an array of the commands' arguments such that it can be
queried using Rust standard library APIs.

Rust does not currently provide a mode to disable this functionality.
For now, then, we need to add an exception.

Bug: 1445935
Change-Id: I9dd6bd76e97d6f8bb33ccee8a7c5b62e0d133e8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4547797
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1150803}
NOKEYCHECK=True
GitOrigin-RevId: 89e2148c00eb2e4d997662a98ba202247c68ca84