Improve flakiness of SSLUITest.TestCloseTabWithUnsafePopup
This test waited for an interstitial but that interstitial is triggered
from a task posted to the message loop that's not counted by the
existing asynchronous navigation waiter.

This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/4553476
except that the test is not actually marked as non-flaky (since it was
still flaky on the bots even with this patch).

The test is removed from the list of Lacros test failures to an ifdef
since this isn't considered a blocker for Lacros.

Bug: 1119359
Change-Id: I545be8963ba94b32d53bdb340f6e7c90154d02ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4581814
Reviewed-by: Carlos IL <carlosil@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1152226}
NOKEYCHECK=True
GitOrigin-RevId: 789e4c3408a135b495dd5533927e672f8d716a4d