[iOS] Autofill Branding refactor
This is an attempt to refactor the existing autofill branding code.

1. Moved files into a separate directory
2. Uses the mediator/coordinator/VC architecture instead of using the
form_input_accessory_view_controller as a sort of coordinator
3. The view BrandingViewController used to be the branding icon. Now it
becomes a wrapper of the icon, so the the view controller could
encapsulate the logic to manage the branding visibility, instead of
taking care of them in the form accessories view controller.
4. Monotone images are removed as we don't support monotone branding
anymore.

The behavior of the branding has also tweaked a bit. Previously, if
autofill becomes not supported DURING the time the keyboard is visible,
the branding would disappear on the spot. As this scenario doesn't
happen (autofill of any type is disabled in Settings, and the user has
to dismiss the keyboard to go there), now the visibility of the branding
is only calculated when the keyboard is about to appear.

Bug: 1447909
Change-Id: Ie5df77094f330172ec7c31da7cb2d2f1afe48cfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4568605
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1154669}
NOKEYCHECK=True
GitOrigin-RevId: 14ddc88ec6bcdfa767186fad16d5eed43a4daa24