Speed up fieldtrial_testing_config presubmit
Overall speed gain (locally) was 12 seconds. As shown below, this gain
will depend a lot on the size of the `out/` directory locally.

The primary speed gain is achieved by introducing a list of
directories to search instead of searching everything in the
project root folder.

A lesser gain (about 2 seconds when testing locally) is achieved by
running the search on a process pool of 4 processes. The bottleneck for
finding all the feature flags is the disk access, but the CPU work of
evaluating the regular expressions do benefit from a process pool.

Bug: 1336663
Change-Id: I1de0785fa736f22b0ec5e5ea91eb6adbd2044d57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4604610
Auto-Submit: Peter Pakkenberg <pbirk@chromium.org>
Commit-Queue: Peter Pakkenberg <pbirk@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1156113}
NOKEYCHECK=True
GitOrigin-RevId: abd6ac397ad96df2902009ac9a93ab97c70254a8