Remove devtools test on Lacros.
The test KeepAliveDevToolsTest.KeepsAliveUntilBrowserClose was disabled
for CHROMEOS_ASH and was failing in Lacros. The tested condition in
chrome/browser/devtools/chrome_devtools_manager_delegate.cc uses the
define IS_CHROMEOS to disable keepalive which was changed in
https://chromium-review.googlesource.com/c/chromium/src/+/3849034

This change makes the test define match the logic, so this test is no
longer compiled for Lacros.

Change-Id: Idbcd1e787f4f2a6496c8f3cc980c932ac2b7350c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4615630
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157683}
NOKEYCHECK=True
GitOrigin-RevId: 86514a3a3be4c87b5199557b4e846aca3cb5a68e