[ios] Fix form controls tests
Previously, we'd used a hard-coded default width and height of 800x600
and this caused scaling issues. The problem was that the tests do not
supply a viewport meta and since the default width was smaller than
the legacy fallback width, it would expand out to 980 and then get
scaled back to 800 to fit on the screen, shrinking the controls.

It's curious that this works on Android, but it seems that on that
platform, we return the native size of the display rather than DIP
(eg, 1440x2170 when emulating a nexus 6), so we wouldn't hit this
issue. Rather than do the same thing and potentially confuse future
readers of the code, we'll supply a larger, test-only size.

Another approach would be to supply the viewport meta for these tests,
but changing the size potentially addresses this issue for a wider
swath of tests and prevents confusing errors in the future.

credit: thanks to bokan@ for pointing me at sources of viewport
scaling and suggesting debugging approaches, msisov@ who suggested
looking at screen ios (which did turn out to be a source of DSF
issues and was useful in the preceding patch -- crrev.com/c/4600124),
and dtapuska@ for info on DIPs and coordinate spaces.

Bug: 1411704,1413706
Change-Id: If74540500278847353d6bb834ce62c4216258312
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4615623
Commit-Queue: Ian Vollick <vollick@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158197}
NOKEYCHECK=True
GitOrigin-RevId: ecc53b773609f908d00bc3758a6b4c0167f2e476