Fix data clearing test on Lacros.
Fix BrowsingDataRemoverBrowserTest.StorageRemovedFromDisk. This test
loads a web page on "localhost", clears the browsing data, and makes
sure that this string doesn't appear in any files in the directory.

Tab restore works differently in Ash vs Lacros. In Ash the default is
that a bubble pops up in the lower-right asking if you want to restore
your tabs, but in Lacros the default is to autmatically restore them.

For this test, the tabs were getting restored, then we were clearing the
browsing data, and then the session restore code was persisting the data
on the still-opened tabs, leading to "localhost" data still being
present. This change just changes the preference to disable tab restore.

Bug: 1455208
Change-Id: I5edd1914276c98c4f12ac1a5975d566f7e148463
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4621590
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1160084}
NOKEYCHECK=True
GitOrigin-RevId: d0155e35e4bccf3314353e0b0e974a1fe31b24ec