[Auto-disable Accessibility] Fix DCHECK edge case and add testing fieldtrial
This CL continues the feature work of auto-disable accessibility.

With this CL we fix a rare edge case that can appear if a user is
rapidly turning accessibility on/off for a single page. We also add the
fieldtrial testing config in preparation for Finch rollout.

AX-Relnotes: N/A
Bug: 1430202, b/265493191
Change-Id: If3f75b5998652595d10db7ce8e6c23cdfd4fdbad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4633740
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Amanda Lin Dietz <aldietz@google.com>
Cr-Commit-Position: refs/heads/main@{#1161215}
NOKEYCHECK=True
GitOrigin-RevId: b6f714835b84899c6136ffb9f50e3be7d922d05f