Remove obsolete/broken functionality from manage.py.
manage.py supports a number of different functions that are no longer
necessary and/or no longer work: writing the json files into a fixed
format, determining which gtests are not swarmed and converting gtests
to be swarmed. All of these operations work directly on the json files,
but those are output by generate_buildbot_json.py, so they already have
a format enforced by the generator and they can't be modified directly
because the pyl files that they are generated from need to be updated
(and dimensions would need to be specified on a per-builder basis).
Checking what tests need to be swarmed could maybe be useful, but it
fails because it attempts to do a git operation on a non-existent
directory, so it's clearly not being used.

The only functionality still being used is the check mode, which checks
that the isolates referenced by builders are defined in
gn_isolate_map.pyl and that the isolates defined in gn_isolate_map.pyl
are referenced by builders. This functionality is being used in
PRESUBMIT.py.

Change-Id: I735a42c33d5f9a0be892211a7c4d37c2145fce98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4639132
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161778}
NOKEYCHECK=True
GitOrigin-RevId: babf03eb61999318d39a5924f985a226d3eb2170