[infra] Remove the need to use $mixin_append.
The $mixin_append field is used for appending rather than overwriting
any of the various *args fields or the named_caches field of the
swarming field, but there are no uses of the *args fields that are not
in $mixin_append and none of the non-$mixin_append occurrences of
named_caches is overwriting any other value. This simplifies the
behavior so that named_caches and *args fields will always be appended
rather than requiring the $mixin_append field. Once all mixins.pyl files
are updated, the $mixin_append handling can be removed.

Bug: 1456553
Change-Id: Ic29508b2b86ac47f3f70472de6e97aadfe6beaf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4632065
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#1161935}
NOKEYCHECK=True
GitOrigin-RevId: 4c35b14c1fb0aeeec6163f6e077d54e88838ba6b