Do not generate can_use_on_swarming_builders field.
With https://crrev.com/c/4637531, the presence of a swarming dict in the
test spec is signal enough to swarm the test and the test spec's
can_use_on_swarming_builders value is not used at all. This change stops
generating the field into the json files. The field is still valid to
include in the the input to the generator script as it provides a means
of being able to disable swarming for an individual test and also
provides the signal that a test is actually intended to be swarmed
rather than just getting a swarming field from a mixin.

Bug: 1268499
Change-Id: I092faa4d21ca0db8d544697c69edde6629f12fc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4641488
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1162626}
NOKEYCHECK=True
GitOrigin-RevId: bb18d53d0eb4e569b7a77d4adc60fc4dc66bfd03