[infra] Remove regen_test_json call from unit test.
The regen_test_json function is intended to simplify the process of
updating the test json files when making changes, but not intended to be
left in when submitted. A use of it was accidentally committed in
crrev.com/c/4632065.

Bug: 1456553
Change-Id: I950fbd480baa600bdb00c2e7ed8b20bca195eb3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4651991
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1163047}
NOKEYCHECK=True
GitOrigin-RevId: 1c1e4d4ba051056022efb98a3c1276a4bb784adc