[GPU] Remove GLImageNativePixmap
Following the deletion of V4L2SliceVDA, nothing in the codebase
references GLImageNativePixmap. This CL deletes it as well as all of the
code that directly references it. Further cleanup is left for followup
CLs.

With this CL, GLImage definitively sails off into the sunset.

Bug: 1310018
Change-Id: I1bf12aa95eaeadf3287b2d37c79ec53e16ce39ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4645468
Reviewed-by: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1164579}
NOKEYCHECK=True
GitOrigin-RevId: 54424945d85153be56bbd06d2618dba5f97a67c5